Make sure zombie processes are killed during package builds
Some of our builds (e.g. kernel) end up leaving more than 13k zombie
processes behind when there's no init process to reap them. We should
use the --as-pid2 option of systemd-nspawn, but it is not present in
our systemd version (219).

To avoid this, we've decided that simply sharing the PID namespace
with the host system is the easiest solution, considering our needs.